+++ docs/CHANGELOG.md
@@ -1,3 +1,1 @@
-* Sync Botania API changes.
-* Support JEI bulk quick transfer.
-* Prevent double-clicking storage UI slots from bulk-transferring stored contents into the inventory.
+* Fix fluid key render